Update NetworkHandler.py
Issue: https://github.com/Yudaotor/EsportsHelper/issues/138

To address this, I implemented a delay and retry logic. This ensures that the earnedDrops endpoint is always captured in the performance log when visiting the reward page. The key changes include adding a retry mechanism with delays, allowing sufficient time for the logs to populate.